# api/openapi.yaml
openapi: 3.1.0
info:
  title: Wuling DevOps API
  version: 0.1.0-stage1
  description: |
    Stage 1 of the Wuling DevOps platform.

    What's implemented:
      - Auth (register, login, /me, personal access tokens, SSH public keys)
      - Orgs and Projects (CRUD)
      - Repositories (CRUD + libgit2-backed read APIs)
      - Git smart HTTP (clone/fetch/push) and Git over SSH (default :2222)
      - Issues (CRUD, labels, comments, assignees)
      - Merge Requests (open/diff/merge with FF/merge-commit/squash, comments, reviews)
      - Wiki (per-project Markdown pages backed by a separate Git repo)
      - Insights (per-project activity, contributors, language stats)

    Planned for later stages (URLs reserved, not yet served):
      - Pipelines

    All errors use the JSON envelope:
      { "error": { "code": "<stable_code>", "message": "<human msg>", "details": {} } }
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /api/v1/orgs/{org_slug}/invitations:
    get:
      summary: List invitations for an org
      parameters:
        - in: path
          name: org_slug
          required: true
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/InvitationStatus" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations:
                    type: array
                    items: { $ref: "#/components/schemas/OrgInvitation" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
    post:
      summary: Create a magic-link invitation
      parameters:
        - in: path
          name: org_slug
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvitationRequest" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitation: { $ref: "#/components/schemas/OrgInvitation" }
                  url:
                    type: string
                    description: Convenience copy of invitation.url.
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "409":
          description: Conflict — pending invitation already exists, or user is already a member.
  /api/v1/orgs/{org_slug}/invitations/{invitation_id}:
    delete:
      summary: Revoke a pending invitation
      parameters:
        - in: path
          name: org_slug
          required: true
          schema: { type: string }
        - in: path
          name: invitation_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: revoked }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /api/v1/invitations/{token}:
    get:
      summary: Preview an invitation by its share token
      description: |
        Returns org name, role, expiry, and status. Identity match is NOT
        enforced on preview so the recipient can see why a link doesn't work
        for them. Acceptance (POST /accept) does enforce the match.
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrgInvitation" }
        "404": { $ref: "#/components/responses/NotFoundError" }
...
  /api/v1/orgs:
    get:
      summary: List orgs the caller belongs to
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  orgs:
                    type: array
                    items: { $ref: "#/components/schemas/Org" }
    post:
      summary: Create a new (non-personal) org; caller becomes its owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug: { type: string, minLength: 2, maxLength: 64 }
                display_name: { type: string, maxLength: 128 }
                description: { type: string, maxLength: 512 }
      responses:
        "201":
          description: created
          content: { application/json: { schema: { $ref: "#/components/schemas/Org" } } }
        "409": { $ref: "#/components/responses/ConflictError" }
  /api/v1/orgs/{org_slug}:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
    get:
      summary: Fetch an org by slug
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Org" } } }
        "404": { $ref: "#/components/responses/NotFoundError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
      description: |
        Used by Git CLI only. The "password" is either a personal access token
        (preferred — recognised by the `wlpat_` prefix) or the user's account
        password.

        **Security:** credentials are sent on every request, so basic auth
        MUST be used over HTTPS in production — configure server URLs as
        `https://` and terminate TLS in front of the API. For local
        development against `localhost` plain HTTP is acceptable.

        Operators should prefer personal access tokens (`wlpat_…`) over
        account passwords, and consider OAuth/bearer-token flows for
        non-CLI clients.
  parameters:
    OrgSlug:
      in: path
      name: org_slug
      required: true
      schema: { type: string }
...
  responses:
...
    ForbiddenError:
      description: caller lacks permission
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFoundError:
      description: resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ConflictError:
      description: resource already exists
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - already_exists
                - rate_limited
                - unsupported
                - bad_request
                - internal
                - unavailable
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Org:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        display_name: { type: string }
        description: { type: string }
        is_personal: { type: boolean }
        created_at: { type: string, format: date-time }
...
    InvitableRole:
      type: string
      enum: [maintainer, developer, reporter, guest]
      description: |
        Subset of OrgRole that can be granted by invitation. Owner promotion
        goes through PATCH /orgs/{slug}/members/{user_id} and is restricted
        to existing owners.
...
    InvitationStatus:
      type: string
      enum: [pending, accepted, revoked, expired]
    OrgInvitation:
      type: object
      properties:
        id:                { type: string, format: uuid }
        org_id:            { type: string, format: uuid }
        org_slug:          { type: string }
        org_display_name:  { type: string }
        inviter:           { $ref: "#/components/schemas/UserRef", nullable: true }
        invitee_user_id:   { type: string, format: uuid, nullable: true }
        invitee_email:     { type: string, nullable: true }
        role:              { $ref: "#/components/schemas/InvitableRole" }
        status:            { $ref: "#/components/schemas/InvitationStatus" }
        expires_at:        { type: string, format: date-time }
        created_at:        { type: string, format: date-time }
        accepted_at:       { type: string, format: date-time, nullable: true }
        token:
          type: string
          description: |
            Raw invitation token. Populated only on the create-invitation
            response — all subsequent reads return the empty string since the
            server stores only the HMAC.
        url:
          type: string
          description: |
            Share URL the inviter should send to the recipient. Populated only
            on the create-invitation response.
    CreateInvitationRequest:
      type: object
      required: [identifier, role]
      properties:
        identifier:
          type: string
          description: |
            Username or email of the invitee. If it contains "@" we treat it
            as an email; otherwise as a username. Existing usernames resolve
            to invitee_user_id; unknown emails are stored as-is and matched
            at accept time.
        role: { $ref: "#/components/schemas/InvitableRole" }
        ttl_hours:
          type: integer
          minimum: 1
          maximum: 720
          description: How long the invitation stays acceptable. Defaults to 168 (7 days).
...
    UserRef:
      type: object
      description: Lightweight user reference embedded in issue/comment payloads.
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        display_name: { type: string }
...